ref js samples
