+++ samples/DTDLParserJSInteropSample/DTDLParserJSInteropSample.csproj
@@ -19,6 +19,7 @@
     <WasmExtraFilesToDeploy Include="index.js" />
 	<WasmExtraFilesToDeploy Include="interfaceInfo.js" />
 	<WasmExtraFilesToDeploy Include="errorInfo.js" />
+	  <WasmExtraFilesToDeploy Include="modelParser.js" />
   </ItemGroup>
 
   <ItemGroup>
